src/main/java/com/odde/massivemailer/controller/InitializePasswordController.java
package com.odde.massivemailer.controller;

import javax.servlet.ServletException;
import javax.servlet.annotation.WebServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

@WebServlet("/initializePassword")
public class InitializePasswordController extends AppController {

    public void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {

        resp.sendRedirect("initialize-success.jsp");
    }
}
